requirements.txt
numpy==1.26.4
torch==2.2.2
torchvision
torchaudio
einops
huggingface-hub>=0.22.0
imageio
matplotlib
ninja
opencv-python
pandas
safetensors
scipy
timm
tqdm
wandb
xformers==0.0.25.post1
neptune
scikit-image
plyfile
torchmetrics
omegaconf
hydra-core
jaxtyping
hydra-submitit-launcher
submitit
GitPython
lightning
ema-pytorch
pytubefix
